Silence warnings in math.hpp

+++ tests/math/ones_digit.cpp
@@ -5,5 +5,5 @@
 TEST(math, ones_digit) {
   EXPECT_EQ(ztl::ones_digit(225), 5);
   EXPECT_EQ(ztl::ones_digit(4238261u), 1u);
-  EXPECT_EQ(ztl::ones_digit(static_cast<uint64_t>(12813968329378591287u)), 7u);
+  EXPECT_EQ(ztl::ones_digit(12813968329378591287ull), 7u);
 }
